refactor(core): split component/mod.rs god-file into submodules
Moves the config/model type definitions out of src/core/component/mod.rs
(1688 lines — baselined GodFile) into new config.rs + model.rs submodules
and extracts the test module to tests.rs, each under structural
thresholds. mod.rs becomes a 53-line re-export root so the
crate::core::component::* import surface is unchanged. Pure move;
baseline row removed.
